Add local GraphQL server for development
Add binary in `dailp-graphql` for running a local version of the GraphQL
server that normally runs in AWS Lambda. This would also theoretically
allow the deployment of a dedicated GraphQL server.

Now developers can run `cargo run --bin dailp-graphql-local` to test it out
