+++ graphql/Cargo.toml
@@ -4,6 +4,14 @@ version = "0.2.0"
 authors = ["Taylor Snead <taylor@snead.xyz>"]
 edition = "2018"
 
+[[bin]]
+name = "dailp-graphql"
+path = "src/lambda.rs"
+
+[[bin]]
+name = "dailp-graphql-local"
+path = "src/server.rs"
+
 [dependencies]
 serde_json = "1.0"
 tokio = { version = "1.5", features = ["full", "time"] }
@@ -16,7 +24,7 @@ dotenv = "0.15"
 lazy_static = "*"
 chrono = "*"
 itertools = "*"
-async_once = {version = "0.1", features = ["tokio"]}
+async_once = "0.2"
 base64 = "*"
 serde = "*"
 serde_with = "1.9"
@@ -24,5 +32,9 @@ serde_plain = "0.3"
 async-trait = "*"
 log = "*"
 pretty_env_logger = "*"
+# Used for running a local server.
+async-graphql-tide = "*"
+tide = "0.16"
+async-std = { version = "1", features = ["attributes", "tokio1"] }
 
 dailp = {path = "../types"}
